+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `RecommenderProtocol.recommend` now expects training data to be provided as a single
   dataframe in experimental representation instead of two separate dataframes in
   computational representation
+- `SubspaceContinuous.samples_random` has been replaced with `SubspaceContinuous.sample`
 
 ### Added
 - `Surrogate` base class now exposes a `to_botorch` method
@@ -36,7 +37,6 @@ _ `_optional` subpackage for managing optional dependencies
 
 ### Deprecations
 - `SequentialGreedyRecommender` class replaced with `BotorchRecommender`
-
 
 ## [0.9.1] - 2024-06-04
 ### Changed